Create USI-LS compatibility
This CFG File gives USI-Life support compatibility to procedural tanks, it is a heavy modification of the TACLS compatibility CFG file.  When the user has USI Life support installed in KSP this will give the player the option to create a tank carrying the different types of supplies for "roverdudes" USI-LS mod.  I have attempted to replicate as close as possible the capacities of "roverdudes" USI-LS parts to ensure that procedurally generated tanks are neither over or under powered.  I have tested the mod extensively and it works with the USI-LS and KSP with issues. please feel free to test