Add KSP 1.3.1 support
Bump mod version to 1.2.13 and tag as compatible with KSP 1.3.1.

Remove Module Manager dll. We depend on it being installed so we do not
need to ship it. Especially with CKAN this is a cleaner approach.

Remove ProceduralParts.dll. This is an output file and should be
included in releases, not in the source repository. Last commits did not
update this file and if you cloned the master branch thinking this dll
would be newer than the release, you would be wrong. Therefore better to
separate sources and build files.

Copy the README to the root directory so that the github page would have
a description.

Credits to cmheisel and HebaruSan for their commits in
https://github.com/Starwaster/ProceduralParts/pull/22
Squashed them and added further changes to rebuild it for KSP 1.3.1.

Note - some paths were hardcoded, which made the rebuild cumbersome.
We should look into how to make the solution file (.sln) portable,
for example using environmental variables.